@@ -3,29 +3,32 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
+    "@emotion/react": "^11.11.4",
+    "@emotion/styled": "^11.11.0",
     "@firebase/app-types": "^0.6.1",
-    "@material-ui/core": "^4.9.7",
-    "@material-ui/icons": "^4.9.1",
+    "@mui/icons-material": "^5.15.14",
+    "@mui/material": "^5.15.14",
+    "@mui/styles": "^5.15.14",
+    "@react-three/fiber": "^8.16.0",
     "bootstrap": "^4.6.0",
     "client-oauth2": "^4.2.5",
     "cookie-parser": "^1.4.5",
-    "firebase": "^8.2.10",
+    "firebase": "^9.0.1",
     "firebase-functions": "^3.11.0",
     "plyr": "^3.7.8",
     "plyr-react": "^5.1.2",
     "querystring": "^0.2.0",
-    "react": "^17.0.1",
+    "react": "^18.0.1",
     "react-async": "^10.0.0",
     "react-beautiful-dnd": "^13.0.0",
-    "react-bootstrap": "^1.0.1",
-    "react-dom": "^17.0.1",
+    "react-bootstrap": "^2.10.2",
+    "react-dom": "^18.2.0",
     "react-firebase-hooks": "^3.0.1",
     "react-player": "^2.12.0",
-    "react-router-dom": "^5.2.0",
+    "react-router-dom": "^6.22.3",
     "react-scripts": "5.0.1",
-    "react-three-fiber": "^5.3.19",
     "react-youtube": "^10.1.0",
-    "three": "^0.126.1"
+    "three": "^0.162.0"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -51,14 +54,13 @@
   "devDependencies": {
     "@types/cookie-parser": "^1.4.2",
     "@types/node": "^14.14.31",
-    "@types/react": "^17.0.2",
-    "@types/react-beautiful-dnd": "^13.0.0",
-    "@types/react-dom": "^17.0.1",
-    "@types/react-router-dom": "^5.1.3",
+    "@types/react-beautiful-dnd": "^13.1.8",
+    "@types/react-dom": "^18.2.23",
     "@types/react-youtube": "^7.6.2",
-    "@types/three": "^0.126.0",
-    "babel-loader": "^9.1.2",
-    "typescript": "^4.2.3",
-    "webpack": "^5.78.0"
+    "@types/three": "^0.162.0",
+    "@typescript-eslint/eslint-plugin": "^7.4.0",
+    "@typescript-eslint/parser": "^7.4.0",
+    "eslint-config-react-app": "^7.0.1",
+    "typescript": "^5.4.3"
   }
 }
