@@ -9,7 +9,8 @@ import {
 } from "react-bootstrap";
 import { AdminToolsContext } from "./AdminToolsProvider";
 import ModalHeader from "react-bootstrap/ModalHeader";
-import firebase from "firebase";
+import firebase from 'firebase/compat/app';
+import 'firebase/compat/database';
 import convertDuration from "./ConvertDuration";
 import { useObjectVal } from "react-firebase-hooks/database";
 
@@ -32,8 +33,8 @@ const AdminToolbox = () => {
 
   return (
     <Modal
-      onShow={() => {}}
-      onHide={() => {}}
+      onShow={() => { }}
+      onHide={() => { }}
       show={toolboxData.video !== null || toolboxData.user !== null}
     >
       <ModalHeader>
