@@ -1,16 +1,16 @@
 import React, { useRef } from 'react'
-import { Canvas, useFrame } from 'react-three-fiber'
-import { Mesh, Material, Group, EdgesGeometry, LineSegments, LineBasicMaterial } from 'three';
+import { Canvas, useFrame } from '@react-three/fiber'
+import { Mesh, Group, EdgesGeometry, LineSegments, LineBasicMaterial, MeshStandardMaterial } from 'three';
 import { STLLoader } from 'three/examples/jsm/loaders/STLLoader';
 
-const NoVideo = (props:any) => {
+const NoVideo = (props: any) => {
   return (
     <>
-    <Canvas>
-      <ambientLight />
-      <Spin />
-    </Canvas>
-    <p className="no-video">No video is currently playing.</p>
+      <Canvas>
+        <ambientLight />
+        <Spin />
+      </Canvas>
+      <p className="no-video">No video is currently playing.</p>
     </>
   )
 
@@ -18,13 +18,13 @@ const NoVideo = (props:any) => {
 
 const Spin = () => {
   // This reference will give us direct access to the mesh
-  const mesh  = useRef<Mesh>();
-  const mat   = useRef<Material>();
-  const group = useRef<Group>();
-  
+  const mesh = useRef<Mesh>(null);
+  const mat = useRef<MeshStandardMaterial>(null);
+  const group = useRef<Group>(null);
+
   // Rotate mesh every frame, this is outside of React without overhead
   useFrame((state, delta) => {
-    if(group?.current !== undefined) {
+    if (group?.current) {
       group.current.rotation.y += delta;
     }
   });
@@ -32,23 +32,23 @@ const Spin = () => {
   new STLLoader().load(
     process.env.PUBLIC_URL + "/amphi.stl",
     (bufferGeometry) => {
-      if(mesh.current === undefined) return;
+      if (!mesh.current) return;
       mesh.current.geometry = bufferGeometry;
 
-      var linesMat = new LineBasicMaterial( { color: 0x343a40 } ) ;
-      var edges    = new EdgesGeometry( bufferGeometry );
-      var lines    = new LineSegments( edges, linesMat );
+      var linesMat = new LineBasicMaterial({ color: 0x343a40 });
+      var edges = new EdgesGeometry(bufferGeometry);
+      var lines = new LineSegments(edges, linesMat);
 
-      if(group.current !== undefined)
+      if (group.current)
         group.current.add(lines);
 
       mesh.current.position.z = -5;
-      lines.position.z        = -5;
+      lines.position.z = -5;
     }
   );
-  
+
   return (
-    <group ref={group} position={[0,0,-100]}>
+    <group ref={group} position={[0, 0, -100]}>
       <mesh ref={mesh}>
         <meshStandardMaterial ref={mat} attach="material" color={0xfdd835} />
       </mesh>
