@@ -1,4 +1,5 @@
-import firebase from "firebase";
+import firebase from 'firebase/compat/app';
+import 'firebase/compat/database';
 import React from "react";
 import { Spinner } from "react-bootstrap";
 import VideoListing from "./VideoListing";
@@ -22,11 +23,11 @@ const Playlist = () => {
   return (
     <>
       {(bucketsArr === null)
-      ? (<p>The global playlist is empty.</p>)
-      :
-      bucketsArr.map((b: any, idx) => (
-        <Bucket bucket={b} key={idx} bucketIdx={idx} />
-      ))
+        ? (<p>The global playlist is empty.</p>)
+        :
+        bucketsArr.map((b: any, idx) => (
+          <Bucket bucket={b} key={idx} bucketIdx={idx} />
+        ))
       }
     </>
   );
