@@ -1,6 +1,8 @@
 import { useAuthState } from "react-firebase-hooks/auth";
 import React from "react";
-import firebase from "firebase";
+import firebase from 'firebase/compat/app';
+import 'firebase/compat/auth';
+import 'firebase/compat/database';
 import { useObjectVal } from "react-firebase-hooks/database";
 
 export type UserState = {
@@ -17,7 +19,7 @@ const noUserState = {
 const UserContext = React.createContext<UserState>(noUserState);
 
 const UserProvider = ({ children }: any) => {
-  const [user, loading, error] = useAuthState(firebase.auth());
+  const [user, , error] = useAuthState(firebase.auth());
   const [udata] = useObjectVal(firebase.database().ref(`users/${user?.uid}`));
 
   return (
